Flytt test til BrevServiceTest og legg til en ekstra test
